+++ lib/abstract_mapper.rb
@@ -11,6 +11,7 @@ require_relative "abstract_mapper/errors/wrong_rule"
 
 require_relative "abstract_mapper/node"
 require_relative "abstract_mapper/branch"
+require_relative "abstract_mapper/command"
 require_relative "abstract_mapper/commands"
 require_relative "abstract_mapper/builder"